finalizando e aplicando a paginação

+++ client/src/pages/Books/index.js
@@ -9,30 +9,31 @@ import './styles.css';
 export default function Books(){
 
     const [books, setBooks] = useState([]);
+    const [page, setPage] = useState(0);
 
     const userName = localStorage.getItem('userName');
 
     const accessToken = localStorage.getItem('accessToken');
-
+    const authorization = {
+        headers: {
+            Authorization: `Bearer ${accessToken}`,
+        }
+    }
     const navigate = useNavigate();
 
     useEffect(()=> {
-        api.get('api/book/v1/asc/20/1', {
-            headers: {
-                Authorization: `Bearer ${accessToken}`,
-            }
-        }).then(response=>{
-            setBooks(response.data.list)
-        })
+        fecthMoreBooks();
     },[accessToken]);
+
+    async function fecthMoreBooks(){
+        const response = await api.get(`api/book/v1/asc/4/${page}`, authorization);
+        setBooks([...books, ...response.data.list]);
+        setPage(page + 1);        
+    }
 
     async function logout(){
         try{
-            await api.get('api/auth/v1/revoke', {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`,
-                }
-            });
+            await api.get('api/auth/v1/revoke', authorization);
 
             localStorage.clear();
             navigate('/');
@@ -44,11 +45,7 @@ export default function Books(){
 
     async function deleteBook(id){
         try{
-            await api.delete(`api/Book/v1/${id}`, {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`,
-                }
-            });
+            await api.delete(`api/Book/v1/${id}`,authorization);
 
             setBooks(books.filter(book => book.id !== id))
         }catch (err){
@@ -99,5 +96,6 @@ export default function Books(){
                 </li>
             ))}
         </ul>
+        <button className="button" onClick={fecthMoreBooks} type="button">Load More</button>
     </div>
 }
